refactor: API 호출 부분을 page 컴포넌트로 이동 후 prop으로 데이터 전달하는 방식으로 변경
